Split ActionConfigurator.py into ActionConfiguratorParts/
